requirements.txt
Flask==3.0.0
pandas==2.1.4
numpy==1.26.2
anndata==0.10.3
mudata==0.2.3
scanpy==1.9.6
scipy==1.11.4
openpyxl==3.1.2
werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
